codebase refactor
