+++ src/components/AdminHeader/AdminHeader.css
@@ -1,31 +1,3 @@
-<template>
-	<main class="admin-header">
-		<div class="card">
-			<div class="card-body pt-2">
-					<div class="row mt-1">
-						<div class="col-4">
-							<div class="input-group mb-3">
-								<div class="input-group-prepend">
-									<div class="input-group-text">
-										<i class="fa fa-search"></i>
-									</div>
-								</div>
-								<input type="text" class="form-control" placeholder="Search">
-							</div>
-						</div>
-					</div>
-			</div>
-		</div>
-	</main>
-</template>
-
-<script>
-export default {
-	name: "admin-header"
-}
-</script>
-
-<style scoped>
 .admin-header .card {
 	border-radius: 0px;
 	height: 4rem;
@@ -58,4 +30,3 @@ export default {
 .admin-header .form-control::placeholder {
 	color: rgba(87, 138, 141, 0.8);
 }
-</style>
