+++ src/components/BarChart.vue
